Use the horizon.cookies functions introduced in Juno
The plugin used jquery.cookies.js functionality directly previously as that
file was included by Horizon itself. Due to refactoring in Juno,
jquery.cookies.js is not directly included anymore -- its functionality
now comes in the form of Python functions with Horizon wrapping around
them. Hence, when loading the Network Topology page in Juno with the previous
versions of the plugin, the only thing that would appear was a Jquery error.

This patch fixes the Contrail dashboard plugin to use the wrappers that are
inclued in Horizon instead of using jquery.cookie.js directly.